# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Plataforma MOOC API
  version: 0.2.0
  description: >
    API REST de la plataforma web de cursos masivos abiertos en línea.
    Prefijo de versión: /api/v1. Errores uniformes ({"error":{"code","message","details"}}),
    sesiones por cookie httpOnly o cabecera Authorization: Bearer <token>.
    Este documento cubre los endpoints implementados en la iteración actual
    (identidad, administración, autoría de cursos, catálogo e inscripciones).
    Pendientes: quizzes, progreso e insignias (ver README).
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /forum/{courseId}:
    get:
      tags: [Foro]
      summary: Hilos del foro de un curso
      description: >
        El foro es del curso, no un tablón público: se lee y se escribe con
        acceso a su contenido —estudiante inscrito, profesor dueño o
        colaborador, administración—. Con `resource_stable_id` se acota a lo
        que se discute sobre una lección.

        Cuelga de /forum y no de /courses/{courseId}/forum por el mismo
        conflicto de patrones que llevó el progreso a
        /enrollments/{courseId}/progress.
      operationId: listThreads
      parameters:
        - { name: courseId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: resource_stable_id, in: query, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Hilos, el más activo primero.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Hilo" }
        "403": { $ref: "#/components/responses/Problema" }
    post:
      tags: [Foro]
      summary: Abre un hilo
      description: >
        Un hilo sobre una lección se ata a su `stable_id` y no a su
        identificador de fila, de modo que la conversación sobrevive a la
        publicación de una versión nueva.
      operationId: createThread
      parameters:
        - { name: courseId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body_md]
              properties:
                title: { type: string }
                body_md: { type: string }
                resource_stable_id: { type: string, format: uuid }
      responses:
        "201":
          description: Hilo abierto.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Hilo" }
  /forum/threads/{threadId}:
    get:
      tags: [Foro]
      summary: Un hilo con sus respuestas
      description: >
        Las respuestas eliminadas se devuelven marcadas y sin cuerpo, no se
        omiten: quitar un mensaje intermedio de una conversación deja sin
        sentido a los que le respondían.
      operationId: getThread
      parameters:
        - { name: threadId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Hilo y respuestas. }
        "403": { $ref: "#/components/responses/Problema" }
  /forum/threads/{threadId}/replies:
    post:
      tags: [Foro]
      summary: Responde a un hilo
      operationId: replyToThread
      parameters:
        - { name: threadId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body_md]
              properties:
                body_md: { type: string }
      responses:
        "201": { description: Respuesta publicada. }
        "409": { description: El hilo está cerrado a nuevas respuestas. }
  /forum/threads/{threadId}/lock:
    post:
      tags: [Foro]
      summary: Cierra o reabre un hilo (solo el profesor del curso)
      operationId: lockThread
      parameters:
        - { name: threadId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { locked: { type: boolean } }
      responses:
        "200": { description: Estado del hilo actualizado. }
  /forum/posts/{postId}:
    delete:
      tags: [Foro]
      summary: Oculta una respuesta
      description: >
        La puede borrar su autor o el profesor del curso; la administración,
        siempre. La respuesta conserva su lugar en la conversación, marcada
        como eliminada y sin cuerpo.
      operationId: deleteForumPost
      parameters:
        - { name: postId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Respuesta ocultada. }
        "403": { $ref: "#/components/responses/Problema" }
...
  /me/export:
    get:
      tags: [Cuenta]
      summary: Exportación de los datos personales de quien la pide
      description: >
        Sin restricción de rol: el derecho a llevarse los propios datos no
        depende de para qué se use la plataforma. El identificador sale de la
        sesión y no de la petición, así que cada quien exporta los suyos y solo
        los suyos.

        No sale lo que no es un dato personal: el hash del token de sesión es
        la llave de la cuenta, y el snapshot de un intento son las preguntas y
        la clave correcta del profesor.

        Se entrega como descarga y con `Cache-Control: no-store`.
      operationId: exportMyData
      responses:
        "200":
          description: Paquete con cuenta, sesiones, inscripciones, progreso, evaluaciones e insignias.
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Problema" }
  /collaborators/{courseId}:
    get:
      tags: [Autoría]
      summary: Coautores de un curso
      description: >
        Solo el dueño (o la administración). Un colaborador edita el curso pero
        no reparte acceso: si pudiera, invitar a alguien equivaldría a cederle
        el curso.

        La ruta cuelga de /collaborators y no de /courses/{courseId}/collaborators
        porque ese prefijo ya tiene /courses/versions/{versionId} y el router
        rechaza en el arranque dos patrones donde un comodín compite con un
        literal en la misma posición.
      operationId: listCollaborators
      parameters:
        - { name: courseId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Coautores del curso.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Colaborador" }
        "403": { $ref: "#/components/responses/Problema" }
    post:
      tags: [Autoría]
      summary: Da acceso de edición a un profesor
      description: >
        Se identifica por correo, que es lo que el dueño tiene a mano. La
        respuesta no distingue "no existe" de "no es profesor": separarlas
        convertiría este endpoint en una forma de averiguar qué correos están
        registrados.
      operationId: addCollaborator
      parameters:
        - { name: courseId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "201": { description: Acceso concedido. }
        "422": { $ref: "#/components/responses/Problema" }
  /collaborators/{courseId}/{userId}:
    delete:
      tags: [Autoría]
      summary: Retira el acceso de un coautor
      operationId: removeCollaborator
      parameters:
        - { name: courseId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Acceso retirado. }
        "404": { $ref: "#/components/responses/Problema" }
...
components:
...
  responses:
...
    Problema:
      description: >
        Error uniforme. El campo `error.code` identifica el motivo de forma
        estable; `error.details` enumera los motivos cuando hay varios, como en
        la validación de publicación.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mooc_session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Hilo:
      type: object
      required: [id, course_id, title, body_md, locked]
      properties:
        id: { type: string, format: uuid }
        course_id: { type: string, format: uuid }
        resource_stable_id:
          type: string
          format: uuid
          description: >
            Ata el hilo a una lección. Es el stable_id, no el identificador de
            fila, para que la conversación sobreviva a una versión nueva.
        author_name: { type: string }
        title: { type: string }
        body_md: { type: string }
        locked: { type: boolean }
        replies: { type: integer }
        created_at: { type: string, format: date-time }
        last_activity_at: { type: string, format: date-time }
    Colaborador:
      type: object
      required: [user_id, email, role]
      properties:
        user_id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
        role: { type: string, enum: [editor] }
        added_at: { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array, items: { type: string } }
...